Adjust skip to allow Windows build on Python 3.5+
VIGRA requires C++11 features, which do not exist reliably until
Microsoft Visual C++ 2015. Based on the Python/Visual C++ compatibility
matrix, this means only Python 3.5 and up can be built with VIGRA. This
leaves out Python 2.7 on Windows. All other Python versions are simply
unsupported by conda-forge. So there is nothing to be done for those
cases. As a result, we still skip Windows builds unless the Python 3.5
minimum is met. This has no effect on Linux or Mac where we build Python
2.7 support on C++11 supporting compilers.

ref: https://wiki.python.org/moin/WindowsCompilers